+++ src/algorithm/nb_iter/jsp/jsp_crtm.cpp
@@ -290,7 +290,7 @@ int JSHOP_cr_trans_mix_Ngbh::prepare_move( int typ )
 
 int JSHOP_cr_trans_mix_Ngbh::do_move()
  {
-   short predJ1;
+   int predJ1;
    int i, help;
    *P[1]=*P[0];
