+++ src/algorithm/nb_iter/jsp/jsp_crtr.cpp
@@ -261,7 +261,7 @@ int JSHOP_cr_trans_Ngbh::prepare_move( int typ )
 
 int JSHOP_cr_trans_Ngbh::do_move()
  {
-   short predJ1;
+   int predJ1;
    int i;
    *P[1]=*P[0];
