+++ src/algorithm/nb_iter/osp/osp_cr_r.cpp
@@ -122,8 +122,8 @@ int OSHOP_cr_TST_Ngbh::prepare_move( int typ )
 
 int OSHOP_cr_TST_Ngbh::do_move()
  {
-   short predJ1;
-   short predM1;
+   int predJ1;
+   int predM1;
    int help;
    *P[1]=*P[0];
    if (witch_swap==JO)
